Error on print statements in tests
Calls to `print` and other output during tests clogs up CI output and
makes it hard to follow. This adds a check for anything written to
stdout or stderr and raises an error if anything is there.

The codebase is not clean with this right now so there is an allowlist
that can be added to test classes: `_ignore_error_on_print_allowlist`.
It can also be enabled locally (it's off by default locally since people
commonly print in tests to debug / develop), with
`PYTORCH_ERROR_ON_TEST_PRINT=1`

ghstack-source-id: e8228be2051b7e3d3b3e1fe74d741be5f8b212b3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62637